// apps/client/src/features/store/components/upgrade/types/upgradeTypes.ts
export interface UpgradePlan {
  _id: string;

  name: string;

  description?: string;

  price: number;

  billingCycle: "MONTHLY" | "YEARLY";

  maxBranches: number;

  maxUsers: number;

  maxProducts: number;

  enableAdvancedReports: boolean;

  enableInventory: boolean;

  enableIntegrations: boolean;

  enableEcommerce: boolean;

  enableInvoiceBranding: boolean;

  prioritySupport: boolean;

  enableMultiLocation: boolean;

  extraFeatures: string[];

  status: "ACTIVE" | "INACTIVE";
}